refactor(test): rename psbt test files to clarify fixed script handling
Files renamed to better indicate their purpose for fixed script compatibility.

Issue: BTC-1845
