+++ packages/wasm-miniscript/test/psbtFixedScriptCompatFixtures.ts
